Move findParentDataset function to useParentDataset hook
