+++ app/scripts/components/exploration/hooks/use-parent-data.ts
@@ -1,8 +1,20 @@
 import { useState, useEffect } from 'react';
 import { useAtom } from 'jotai';
 import { externalDatasetsAtom } from '../atoms/datasetLayers';
-import { findParentDataset } from '$utils/data-utils';
 import { DatasetData } from '$types/veda';
+
+function findParentDataset({
+  datasets,
+  datasetId
+}: {
+  datasets: DatasetData[];
+  datasetId: string;
+}): DatasetData | undefined {
+  const parentDatset = datasets.find(
+    (dataset: DatasetData) => dataset.id === datasetId
+  );
+  return parentDatset;
+}
 
 export default function useParentDataset({ datasetId }: { datasetId: string }) {
   const [currentDataset, setCurrentDataset] = useState<DatasetData | undefined>(
